+++ devicemanagement/tests/TestDevMgmtApiSyncCmds.h
@@ -145,7 +145,6 @@ protected:
   // Security tests thermal and power management service
   void setModuleActivePowerManagementRange(bool singleDevice);
   void setModuleSetTemperatureThresholdRange(bool singleDevice);
-  void setModuleStaticTDPLevelRange(bool singleDevice);
   void setModuleActivePowerManagementRangeInvalidInputSize(bool singleDevice);
   void setModuleActivePowerManagementRangeInvalidInputBuffer(bool singleDevice);